+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Include openEO API error id automatically in exception message to simplify user support and post-mortem analysis.
+- Use `Connection.default_timeout` (when set) also on version discovery request
 
 
 ### Removed